Code review changes - Part 4
